chore: refatoracao do codigo e validacoes de campos
